+++ src/Selah.Infrastructure/Repository/Interfaces/IBankingRepository.cs
@@ -11,20 +11,20 @@ namespace Selah.Infrastructure.Repository.Interfaces
         /// </summary>
         /// <param name="userId"></param>
         /// <returns></returns>
-        public Task<IEnumerable<BankAccountSql>> GetAccounts(int userId, int limit, int offset);
+        public Task<IEnumerable<BankAccountSql>> GetAccounts(long userId, int limit, int offset);
 
         /// <summary>
         /// Returns a list of accounts by institution
         /// </summary>
         /// <param name="institutionId"></param>
         /// <returns></returns>
-        public Task<IEnumerable<BankAccountSql>> GetAccountsByInstitutionId(int institutionId);
+        public Task<IEnumerable<BankAccountSql>> GetAccountsByInstitutionId(long institutionId);
 
         /// <summary>
         /// Returns a bank account by Id
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public Task<BankAccountSql> GetAccountById(int id);
+        public Task<BankAccountSql> GetAccountById(long id);
     }
 }
